+++ CHANGELOG.md
@@ -17,6 +17,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Breaking changes
 
+* Deprecated `ui.always-allow-large-revsets` setting has been removed.
+
 ### Deprecations
 
 * The revset function `diff_contains()` has been renamed to `diff_lines()`.